configuration/requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
gunicorn==20.1.0
mccabe==0.7.0
pep8-naming==0.13.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3
